マイグレーションの問題を改善

+++ templates/Admin/PermissionGroups/edit.php
@@ -56,18 +56,32 @@ $this->BcAdmin->addAdminMainBodyHeaderLinks([
         'data-bca-btn-width' => 'lg',
         'id' => 'BtnSave']
     ) ?>
+    <?php echo $this->BcHtml->link(__d('baser_core', 'アクセスルール一覧に移動'), [
+      'controller' => 'Permissions',
+      'action' => 'index',
+      $userGroupId,
+      '?' => [
+        'permission_group_type' => $entity->type,
+        'permission_group_id' => $entity->id,
+      ],
+    ], [
+      'class' => 'bca-btn bca-actions__item',
+    ]) ?>
   </div>
   <div class="bca-actions__sub">
-      <?= $this->BcAdminForm->postLink(
-        __d('baser_core', '削除'),
-        ['action' => 'delete', $userGroupId, $entity->id],
-        ['block' => true,
-          'confirm' => __d('baser_core', '{0} を本当に削除してもいいですか？', $entity->name),
-          'class' => 'bca-submit-token button bca-btn bca-actions__item',
-          'data-bca-btn-type' => 'delete',
-          'data-bca-btn-size' => 'sm'
-        ]
-      ) ?>
+      <?php if ($isDeletable): ?>
+        <?= $this->BcAdminForm->postLink(
+          __d('baser_core', '削除'),
+          ['action' => 'delete', $userGroupId, $entity->id],
+          ['block' => true,
+            'confirm' => __d('baser_core', '{0} を本当に削除してもいいですか？', $entity->name),
+            'class' => 'bca-submit-token button bca-btn bca-actions__item',
+            'data-bca-btn-type' => 'delete',
+            'data-bca-btn-size' => 'sm',
+            'data-bca-btn-color' => "danger"
+          ]
+        ) ?>
+      <?php endif ?>
   </div>
 </div>
